refactor : moving manufacturer to folder entities
