+++ src/domain/entities/manufacturer.ts
