+++ src/domain/repositories/manufacturer-repository.ts
@@ -1,4 +1,4 @@
-import { Manufacturer } from "../manufacturer";
+import { Manufacturer } from "../entities/manufacturer";
 
 export interface AddManufacturerRepository {
   create(data: Manufacturer): Promise<Manufacturer>;
