+++ src/domain/usecases/add-manufacturer-use-case.ts
@@ -1,4 +1,4 @@
-import { Manufacturer } from "../manufacturer";
+import { Manufacturer } from "../entities/manufacturer";
 
 export interface AddManufacturer {
   execute(data: Manufacturer): Promise<Manufacturer>;
